+++ tests/integration/proxy_test.go
@@ -446,6 +446,7 @@ func setupReadOnlyProxyCluster(t *testing.T, proxyArgs ...string) *readOnlyProxy
 	initialClusterSpec := &cluster.ClusterSpec{
 		InitMode:               cluster.ClusterInitModeP(cluster.ClusterInitModeNew),
 		SleepInterval:          &cluster.Duration{Duration: 2 * time.Second},
+		RequestTimeout:         &cluster.Duration{Duration: 1 * time.Second},
 		FailInterval:           &cluster.Duration{Duration: 5 * time.Second},
 		ConvergenceTimeout:     &cluster.Duration{Duration: 30 * time.Second},
 		MaxStandbyLag:          cluster.Uint32P(50 * 1024),